backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==12.0
pydantic>=2.0.0
email-validator>=2.0.0
